SafeRoute_AI/requirements.txt
fastapi
uvicorn
supabase
osmnx
networkx
pandas
python-dotenv
supabase
twilio 
shapely 
pydantic
scikit-learn